@@ -3,18 +3,18 @@
   "version": "1.0.0",
   "description": "A GitHub App that checks pull requests for ruleset compliance and comments if bypassed.",
   "main": "src/app.js",
+  "type": "module",
   "scripts": {
     "start": "node src/app.js",
     "dev": "nodemon src/app.js",
     "dev:fetch": "NODE_OPTIONS='--experimental-fetch' nodemon src/app.js",
-    "smee": "smee -u ${WEBHOOK_PROXY_URL:-https://smee.io/new} -t http://localhost:${PORT:-3000}/webhook",
-    "smee:new": "smee -u -t http://localhost:${PORT:-3000}/webhook",
-    "test": "jest",
+    "smee": "npx smee-client -u ${WEBHOOK_PROXY_URL} -t http://localhost:${PORT:-3000}/webhook",
+    "smee:new": "npx smee-client --generate-unique-url -t http://localhost:${PORT:-3000}/webhook",
+    "test": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --no-cache",
     "lint": "eslint .",
     "docker:build": "docker build -t github-pr-bypass-checker .",
     "docker:run": "docker run -p ${PORT:-3000}:3000 --env-file .env github-pr-bypass-checker",
-    "docker:prod": "docker build -t github-pr-bypass-checker:production . && docker run -p ${PORT:-3000}:3000 -e NODE_ENV=production --env-file .env.production github-pr-bypass-checker:production",
-    "prepare": "husky install || echo 'Skipping husky installation'"
+    "docker:prod": "docker build -t github-pr-bypass-checker:production . && docker run -p ${PORT:-3000}:3000 -e NODE_ENV=production --env-file .env.production github-pr-bypass-checker:production"
   },
   "dependencies": {
     "@octokit/auth-app": "^7.1.5",
@@ -23,19 +23,22 @@
     "body-parser": "^2.2.0",
     "dotenv": "^16.0.3",
     "express": "^5.1.0",
-    "node-fetch": "^3.3.2",
-    "octokit": "^4.1.2",
-    "smee-client": "^3.1.1"
+    "octokit": "^4.1.2"
   },
   "devDependencies": {
-    "eslint": "^9.23.0",
-    "husky": "^9.1.7",
+    "@eslint/js": "^9.24.0",
+    "cross-env": "^7.0.3",
+    "eslint": "^9.24.0",
+    "eslint-plugin-jest": "^28.11.0",
+    "globals": "^16.0.0",
     "jest": "^29.4.3",
+    "node-fetch": "^2.7.0",
     "nodemon": "^3.1.9",
+    "smee-client": "^3.1.1",
     "supertest": "^7.1.0"
   },
   "engines": {
-    "node": ">=16"
+    "node": ">=18"
   },
   "keywords": [
     "github",
