@@ -1,6 +1,5 @@
-const { createOctokitClient, postComment, checkRepoBypassedRuleSuites, checkOrgBypassedRulesSuites } = require('../utils/github');
-const logger = require('../utils/logger');
-
+import { createOctokitClient, postComment, checkRepoBypassedRuleSuites } from '../utils/github.js';
+import logger from '../utils/logger.js';
 /**
  * Validate if a ruleset is being bypassed
  * @param {Object} pullRequest - Pull request object from webhook payload
@@ -70,6 +69,10 @@ async function handlePullRequestClosed(pullRequest) {
         
         // Get authenticated Octokit client
         const octokit = await createOctokitClient();
+        if (!octokit) {
+            logger.error('Failed to create Octokit client, aborting ruleset check');
+            return;
+        }
         
         // Initialize variables to collect bypass information
         let repoBypassedRuleSuites = [];
@@ -103,6 +106,7 @@ async function handlePullRequestClosed(pullRequest) {
         }
     } catch (error) {
         logger.error(`Error processing closed pull request: ${error.message}`);
+        logger.debug(error.stack);
     }
 }
 
@@ -151,7 +155,7 @@ function formatRuleSuites(ruleSuites) {
         return '';
     }
     
-    return ruleSuites.map((ruleSuite, index) => {
+    return ruleSuites.map((ruleSuite) => {
         // Extract ruleset information with fallbacks for different API response formats
         const beforeSha = ruleSuite.before_sha ? ruleSuite.before_sha.substring(0, 7) : 'Unknown';
         const afterSha = ruleSuite.after_sha ? ruleSuite.after_sha.substring(0, 7) : 'Unknown';
@@ -166,4 +170,7 @@ function formatRuleSuites(ruleSuites) {
     }).join('\n\n');
 }
 
-module.exports = { handlePullRequest, validateRuleset };
+export { 
+    handlePullRequest, 
+    validateRuleset 
+};
